Fix Bluetooth detection on macOS post-High-Sierra
If we detect macOS, we now switch to "noble-mac" instead of the regular
"noble". "noble-mac" is a port of "noble", but uses Core Bluetooth
instead of trying to directly communicate with macOS's Bluetooth stack
via XPC ... which tends to break with new macOS releases, as it's
considered an implementation detail.